pointer and temp segments
